pyproject.toml
[project]
name = "regenova-intel"
version = "0.1.0"
description = "REGENOVA-Intel: AI-powered peptide prescribing assistant with evidence-tiered retrieval"
authors = [
    { name = "zoe786", email = "your@email.com" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.8",
    "langchain-community>=0.2.0",
    "chromadb>=0.5.0",
    "networkx>=3.3",
    "biopython>=1.83",
    "youtube-transcript-api>=0.6.2",
    "beautifulsoup4>=4.12.3",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "streamlit>=1.35.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "rich>=13.7.1",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["apps", "pipelines", "knowledge", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "pipelines", "knowledge"]
